+++ packages/core/src/index.ts
@@ -3,8 +3,10 @@
  * It contains the `createEnv` function that you can use to create your schema.
  * @module
  */
-import type { TypeOf, ZodError, ZodObject, ZodType } from "zod";
-import { object } from "zod";
+import type { StandardSchemaDictionary, StandardSchemaV1 } from "./standard";
+import { ensureSynchronous, parseWithDictionary } from "./standard";
+
+export type { StandardSchemaV1, StandardSchemaDictionary };
 
 /** @internal */
 export type ErrorMessage<T extends string> = T;
@@ -13,7 +15,14 @@ export type Simplify<T> = {
   [P in keyof T]: T[P];
 } & {};
 
-// biome-ignore lint/suspicious/noExplicitAny: accept any object
+type PossiblyUndefinedKeys<T> = {
+  [K in keyof T]: undefined extends T[K] ? K : never;
+}[keyof T];
+
+type UndefinedOptional<T> = Partial<Pick<T, PossiblyUndefinedKeys<T>>> &
+  Omit<T, PossiblyUndefinedKeys<T>>;
+
+// biome-ignore lint/suspicious/noExplicitAny: <explanation>
 type Impossible<T extends Record<string, any>> = Partial<
   Record<keyof T, never>
 >;
@@ -21,13 +30,13 @@ type Impossible<T extends Record<string, any>> = Partial<
 type UnReadonlyObject<T> = T extends Readonly<infer U> ? U : T;
 
 type Reduce<
-  TArr extends Array<Record<string, unknown>>,
-  TAcc = {},
+  TArr extends Record<string, unknown>[],
+  TAcc = object,
 > = TArr extends []
   ? TAcc
   : TArr extends [infer Head, ...infer Tail]
-    ? Tail extends Array<Record<string, unknown>>
-      ? Head & Reduce<Tail, TAcc>
+    ? Tail extends Record<string, unknown>[]
+      ? UnReadonlyObject<Head> & Omit<Reduce<Tail, TAcc>, keyof Head>
       : never
     : never;
 
@@ -35,7 +44,7 @@ type Reduce<
  * The options that can be passed to the `createEnv` function.
  */
 export interface BaseOptions<
-  TShared extends Record<string, ZodType>,
+  TShared extends StandardSchemaDictionary,
   TExtends extends Array<Record<string, unknown>>,
 > {
   /**
@@ -59,7 +68,7 @@ export interface BaseOptions<
    * Called when validation fails. By default the error is logged,
    * and an error is thrown telling what environment variables are invalid.
    */
-  onValidationError?: (error: ZodError) => never;
+  onValidationError?: (issues: readonly StandardSchemaV1.Issue[]) => never;
 
   /**
    * Called when a server-side environment variable is accessed on the client.
@@ -95,7 +104,7 @@ export interface BaseOptions<
  * your framework performs static analysis and tree-shakes unused variables.
  */
 export interface LooseOptions<
-  TShared extends Record<string, ZodType>,
+  TShared extends StandardSchemaDictionary,
   TExtends extends Array<Record<string, unknown>>,
 > extends BaseOptions<TShared, TExtends> {
   runtimeEnvStrict?: never;
@@ -116,9 +125,9 @@ export interface LooseOptions<
  */
 export interface StrictOptions<
   TPrefix extends string | undefined,
-  TServer extends Record<string, ZodType>,
-  TClient extends Record<string, ZodType>,
-  TShared extends Record<string, ZodType>,
+  TServer extends StandardSchemaDictionary,
+  TClient extends StandardSchemaDictionary,
+  TShared extends StandardSchemaDictionary,
   TExtends extends Array<Record<string, unknown>>,
 > extends BaseOptions<TShared, TExtends> {
   /**
@@ -156,7 +165,7 @@ export interface StrictOptions<
  */
 export interface ClientOptions<
   TPrefix extends string | undefined,
-  TClient extends Record<string, ZodType>,
+  TClient extends StandardSchemaDictionary,
 > {
   /**
    * The prefix that client-side variables must have. This is enforced both at
@@ -183,7 +192,7 @@ export interface ClientOptions<
  */
 export interface ServerOptions<
   TPrefix extends string | undefined,
-  TServer extends Record<string, ZodType>,
+  TServer extends StandardSchemaDictionary,
 > {
   /**
    * Specify your server-side environment variables schema here. This way you can ensure the app isn't
@@ -202,10 +211,26 @@ export interface ServerOptions<
   }>;
 }
 
+export interface CreateSchemaOptions<
+  TServer extends StandardSchemaDictionary,
+  TClient extends StandardSchemaDictionary,
+  TShared extends StandardSchemaDictionary,
+  TFinalSchema extends StandardSchemaV1<{}, {}>,
+> {
+  /**
+   * A custom function to combine the schemas.
+   * Can be used to add further refinement or transformation.
+   */
+  createFinalSchema?: (
+    shape: TServer & TClient & TShared,
+    isServer: boolean,
+  ) => TFinalSchema;
+}
+
 export type ServerClientOptions<
   TPrefix extends string | undefined,
-  TServer extends Record<string, ZodType>,
-  TClient extends Record<string, ZodType>,
+  TServer extends StandardSchemaDictionary,
+  TClient extends StandardSchemaDictionary,
 > =
   | (ClientOptions<TPrefix, TClient> & ServerOptions<TPrefix, TServer>)
   | (ServerOptions<TPrefix, TServer> & Impossible<ClientOptions<never, never>>)
@@ -213,37 +238,41 @@ export type ServerClientOptions<
 
 export type EnvOptions<
   TPrefix extends string | undefined,
-  TServer extends Record<string, ZodType>,
-  TClient extends Record<string, ZodType>,
-  TShared extends Record<string, ZodType>,
+  TServer extends StandardSchemaDictionary,
+  TClient extends StandardSchemaDictionary,
+  TShared extends StandardSchemaDictionary,
   TExtends extends Array<Record<string, unknown>>,
-> =
+  TFinalSchema extends StandardSchemaV1<{}, {}>,
+> = (
   | (LooseOptions<TShared, TExtends> &
       ServerClientOptions<TPrefix, TServer, TClient>)
   | (StrictOptions<TPrefix, TServer, TClient, TShared, TExtends> &
-      ServerClientOptions<TPrefix, TServer, TClient>);
+      ServerClientOptions<TPrefix, TServer, TClient>)
+) &
+  CreateSchemaOptions<TServer, TClient, TShared, TFinalSchema>;
 
 type TPrefixFormat = string | undefined;
-type TServerFormat = Record<string, ZodType>;
-type TClientFormat = Record<string, ZodType>;
-type TSharedFormat = Record<string, ZodType>;
+type TServerFormat = StandardSchemaDictionary;
+type TClientFormat = StandardSchemaDictionary;
+type TSharedFormat = StandardSchemaDictionary;
 type TExtendsFormat = Array<Record<string, unknown>>;
 
-/**
- * Creates a new environment variable schema.
- */
-export type CreateEnv<
+export type DefaultCombinedSchema<
   TServer extends TServerFormat,
   TClient extends TClientFormat,
   TShared extends TSharedFormat,
+> = StandardSchemaV1<
+  {},
+  UndefinedOptional<
+    StandardSchemaDictionary.InferOutput<TServer & TClient & TShared>
+  >
+>;
+
+export type CreateEnv<
+  TFinalSchema extends StandardSchemaV1<{}, {}>,
   TExtends extends TExtendsFormat,
 > = Readonly<
-  Simplify<
-    TypeOf<ZodObject<TServer>> &
-      TypeOf<ZodObject<TClient>> &
-      TypeOf<ZodObject<TShared>> &
-      UnReadonlyObject<Reduce<TExtends>>
-  >
+  Simplify<Reduce<[StandardSchemaV1.InferOutput<TFinalSchema>, ...TExtends]>>
 >;
 
 /**
@@ -255,9 +284,14 @@ export function createEnv<
   TClient extends TClientFormat = NonNullable<unknown>,
   TShared extends TSharedFormat = NonNullable<unknown>,
   const TExtends extends TExtendsFormat = [],
+  TFinalSchema extends StandardSchemaV1<{}, {}> = DefaultCombinedSchema<
+    TServer,
+    TClient,
+    TShared
+  >,
 >(
-  opts: EnvOptions<TPrefix, TServer, TClient, TShared, TExtends>,
-): CreateEnv<TServer, TClient, TShared, TExtends> {
+  opts: EnvOptions<TPrefix, TServer, TClient, TShared, TExtends, TFinalSchema>,
+): CreateEnv<TFinalSchema, TExtends> {
   const runtimeEnv = opts.runtimeEnvStrict ?? opts.runtimeEnv ?? process.env;
 
   const emptyStringAsUndefined = opts.emptyStringAsUndefined ?? false;
@@ -276,43 +310,50 @@ export function createEnv<
   const _client = typeof opts.client === "object" ? opts.client : {};
   const _server = typeof opts.server === "object" ? opts.server : {};
   const _shared = typeof opts.shared === "object" ? opts.shared : {};
-  const client = object(_client);
-  const server = object(_server);
-  const shared = object(_shared);
   const isServer =
     opts.isServer ?? (typeof window === "undefined" || "Deno" in window);
 
-  const allClient = client.merge(shared);
-  const allServer = server.merge(shared).merge(client);
-  const parsed = isServer
-    ? allServer.safeParse(runtimeEnv) // on server we can validate all env vars
-    : allClient.safeParse(runtimeEnv); // on client we can only validate the ones that are exposed
+  const finalSchemaShape = isServer
+    ? {
+        ..._server,
+        ..._shared,
+        ..._client,
+      }
+    : {
+        ..._client,
+        ..._shared,
+      };
+
+  const parsed =
+    opts
+      .createFinalSchema?.(finalSchemaShape as never, isServer)
+      ["~standard"].validate(runtimeEnv) ??
+    parseWithDictionary(finalSchemaShape, runtimeEnv);
+
+  ensureSynchronous(parsed, "Validation must be synchronous");
 
   const onValidationError =
     opts.onValidationError ??
-    ((error: ZodError) => {
-      console.error(
-        "❌ Invalid environment variables:",
-        error.flatten().fieldErrors,
-      );
+    ((issues) => {
+      console.error("❌ Invalid environment variables:", issues);
       throw new Error("Invalid environment variables");
     });
 
   const onInvalidAccess =
     opts.onInvalidAccess ??
-    ((_variable: string) => {
+    (() => {
       throw new Error(
         "❌ Attempted to access a server-side environment variable on the client",
       );
     });
 
-  if (parsed.success === false) {
-    return onValidationError(parsed.error);
+  if (parsed.issues) {
+    return onValidationError(parsed.issues);
   }
 
   const isServerAccess = (prop: string) => {
     if (!opts.clientPrefix) return true;
-    return !prop.startsWith(opts.clientPrefix) && !(prop in shared.shape);
+    return !prop.startsWith(opts.clientPrefix) && !(prop in _shared);
   };
   const isValidServerAccess = (prop: string) => {
     return isServer || !isServerAccess(prop);
@@ -324,7 +365,7 @@ export function createEnv<
   const extendedObj = (opts.extends ?? []).reduce((acc, curr) => {
     return Object.assign(acc, curr);
   }, {});
-  const fullObj = Object.assign(parsed.data, extendedObj);
+  const fullObj = Object.assign(extendedObj, parsed.value);
 
   const env = new Proxy(fullObj, {
     get(target, prop) {