+++ packages/nuxt/src/index.ts
@@ -5,23 +5,28 @@
  */
 import type {
   CreateEnv,
+  CreateSchemaOptions,
+  DefaultCombinedSchema,
   ServerClientOptions,
+  StandardSchemaDictionary,
+  StandardSchemaV1,
   StrictOptions,
 } from "@t3-oss/env-core";
 import { createEnv as createEnvCore } from "@t3-oss/env-core";
-import type { ZodType } from "zod";
 
 const CLIENT_PREFIX = "NUXT_PUBLIC_" as const;
 type ClientPrefix = typeof CLIENT_PREFIX;
 
 type Options<
-  TServer extends Record<string, ZodType>,
-  TClient extends Record<`${ClientPrefix}${string}`, ZodType>,
-  TShared extends Record<string, ZodType>,
+  TServer extends StandardSchemaDictionary,
+  TClient extends Record<`${ClientPrefix}${string}`, StandardSchemaV1>,
+  TShared extends StandardSchemaDictionary,
   TExtends extends Array<Record<string, unknown>>,
+  TFinalSchema extends StandardSchemaV1<{}, {}>,
 > = Omit<
   StrictOptions<ClientPrefix, TServer, TClient, TShared, TExtends> &
-    ServerClientOptions<ClientPrefix, TServer, TClient>,
+    ServerClientOptions<ClientPrefix, TServer, TClient> &
+    CreateSchemaOptions<TServer, TClient, TShared, TFinalSchema>,
   "runtimeEnvStrict" | "runtimeEnv" | "clientPrefix"
 >;
 
@@ -29,21 +34,33 @@ type Options<
  * Create a new environment variable schema.
  */
 export function createEnv<
-  TServer extends Record<string, ZodType> = NonNullable<unknown>,
+  TServer extends StandardSchemaDictionary = NonNullable<unknown>,
   TClient extends Record<
     `${ClientPrefix}${string}`,
-    ZodType
+    StandardSchemaV1
   > = NonNullable<unknown>,
-  TShared extends Record<string, ZodType> = NonNullable<unknown>,
+  TShared extends StandardSchemaDictionary = NonNullable<unknown>,
   const TExtends extends Array<Record<string, unknown>> = [],
+  TFinalSchema extends StandardSchemaV1<{}, {}> = DefaultCombinedSchema<
+    TServer,
+    TClient,
+    TShared
+  >,
 >(
-  opts: Options<TServer, TClient, TShared, TExtends>,
-): CreateEnv<TServer, TClient, TShared, TExtends> {
+  opts: Options<TServer, TClient, TShared, TExtends, TFinalSchema>,
+): CreateEnv<TFinalSchema, TExtends> {
   const client = typeof opts.client === "object" ? opts.client : {};
   const server = typeof opts.server === "object" ? opts.server : {};
   const shared = opts.shared;
 
-  return createEnvCore<ClientPrefix, TServer, TClient, TShared, TExtends>({
+  return createEnvCore<
+    ClientPrefix,
+    TServer,
+    TClient,
+    TShared,
+    TExtends,
+    TFinalSchema
+  >({
     ...opts,
     shared,
     client,